requirements.txt
aiofiles==23.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiostream==0.4.5
annotated-types==0.6.0
anthropic==0.3.6
anyio==3.6.2
appdirs==1.4.4
argilla==0.0.1
arxiv==1.4.7
asgi-csrf==0.9
asgiref==3.6.0
async-timeout==4.0.2
asyncer==0.0.2
attrs==23.1.0
auth0-python==4.4.0
Authlib==1.2.1
backoff==2.2.1
beautifulsoup4==4.12.2
bidict==0.22.1
bleach==6.0.0
blinker==1.6.2
buildozer==1.5.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chainlit==0.6.1
chardet==5.1.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
click-default-group-wheel==1.2.2
click-log==0.4.0
clickhouse-connect==0.6.8
cmake==3.27.0
cohere==4.18.0
colorama==0.4.6
coloredlogs==15.0.1
cryptography==41.0.2
ctransformers==0.2.14
dataclasses-json==0.5.13
datasette==0.64.3
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
dill==0.3.7
diskcache==5.6.1
distlib==0.3.6
distro==1.8.0
dnspython==2.4.1
docarray==0.21.1
docker==6.1.3
docstring-parser==0.15
docutils==0.20.1
dotty-dict==1.3.1
duckdb==0.8.1
ecdsa==0.18.0
et-xmlfile==1.1.0
exceptiongroup==1.1.1
faiss-cpu==1.7.4
fake-useragent==1.1.3
fastapi==0.98.0
fastapi-socketio==0.0.10
fastavro==1.7.4
feedparser==6.0.10
filelock==3.12.0
filetype==1.2.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-Testing==0.8.1
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.6.0
geographiclib==2.0
geopy==2.3.0
gitdb==4.0.10
GitPython==3.1.32
google-api-core==2.11.1
google-api-python-client==2.95.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.1
gotrue==1.0.2
gpt4all==1.0.5
greenlet==2.0.2
grpcio==1.56.2
grpcio-health-checking==1.56.2
grpcio-reflection==1.56.2
grpcio-tools==1.56.2
gunicorn==20.1.0
h11==0.14.0
h2==4.1.0
hnswlib==0.7.0
hpack==4.0.0
httpcore==0.16.3
httplib2==0.22.0
httptools==0.6.0
httpx==0.23.3
huggingface-hub==0.15.1
humanfriendly==10.0
hupper==1.12
hyperframe==6.0.1
idna==3.4
importlib-metadata==6.0.1
iniconfig==2.0.0
instructor==0.2.9
invoke==1.7.3
itsdangerous==2.1.2
janus==1.0.0
jaraco.classes==3.3.0
jcloud==0.2.13
jeepney==0.8.0
jina==3.15.2
jina-hubble-sdk==0.39.0
Jinja2==3.1.2
joblib==1.3.1
keyring==24.2.0
langchain==0.0.219
langchainplus-sdk==0.0.20
langflow==0.2.7
lit==16.0.6
litechain==0.1.2
llama-cpp-python==0.1.77
loguru==0.7.0
lxml==4.9.3
lz4==4.3.2
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.2
marshmallow==3.20.1
mdurl==0.1.2
mergedeep==1.3.4
monotonic==1.6
more-itertools==10.0.0
mpmath==1.3.0
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==3.1
nltk==3.8.1
nodeenv==1.8.0
numexpr==2.8.4
numpy==1.25.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
olefile==0.46
onnxruntime==1.15.1
openai==0.28.1
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
opentelemetry-api==1.18.0
opentelemetry-exporter-otlp==1.18.0
opentelemetry-exporter-otlp-proto-common==1.18.0
opentelemetry-exporter-otlp-proto-grpc==1.18.0
opentelemetry-exporter-otlp-proto-http==1.18.0
opentelemetry-exporter-prometheus==1.12.0rc1
opentelemetry-instrumentation==0.39b0
opentelemetry-instrumentation-aiohttp-client==0.39b0
opentelemetry-instrumentation-asgi==0.39b0
opentelemetry-instrumentation-fastapi==0.39b0
opentelemetry-instrumentation-grpc==0.39b0
opentelemetry-proto==1.18.0
opentelemetry-sdk==1.18.0
opentelemetry-semantic-conventions==0.39b0
opentelemetry-util-http==0.39b0
orjson==3.9.2
overrides==7.3.1
packaging==23.1
pandas==2.0.3
pathspec==0.11.1
pdf2image==1.16.3
pdfminer.six==20221105
pdoc3==0.10.0
pexpect==4.8.0
Pillow==10.0.0
pinecone-client==2.2.2
Pint==0.21
pkginfo==1.9.6
platformdirs==3.5.1
pluggy==1.0.0
portalocker==2.7.0
postgrest==0.10.6
posthog==3.0.1
prisma==0.9.1
prometheus-client==0.17.1
protobuf==4.23.4
psycopg==3.1.9
psycopg-binary==3.1.9
psycopg2-binary==2.9.6
ptyprocess==0.7.0
pulsar-client==3.2.0
py==1.11.0
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.15.1
PyJWT==2.8.0
pymongo==4.4.1
pypandoc==1.11
pyparsing==3.1.0
pypdf==3.13.0
pysrt==1.1.2
pytest==7.3.1
pytest-mock==3.10.0
pytest-timeout==2.1.0
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-engineio==4.5.1
python-gitlab==3.15.0
python-graphql-client==0.4.3
python-jose==3.3.0
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-semantic-release==7.33.2
python-socketio==5.8.0
pytz==2023.3
PyYAML==6.0
qdrant-client==1.3.1
readme-renderer==40.0
realtime==1.0.0
regex==2023.6.3
requests==2.30.0
requests-toolbelt==1.0.0
retry==0.9.2
rfc3986==1.5.0
rich==13.4.2
rsa==4.9
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
SecretStorage==3.3.3
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.99
sgmllib3k==1.0.0
sh==2.0.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==1.4.41
sqlalchemy2-stubs==0.0.2a35
sqlmodel==0.0.8
starlette==0.27.0
storage3==0.5.3
StrEnum==0.4.15
supabase==1.0.3
supafunc==0.2.2
sympy==1.12
syncer==2.0.3
tabulate==0.9.0
tenacity==8.2.2
threadpoolctl==3.2.0
tiktoken==0.4.0
tokenizers==0.13.3
tomli==2.0.1
tomlkit==0.11.8
torch==2.0.1
torchvision==0.15.2
tqdm==4.65.0
transformers==4.31.0
triton==2.0.0
twine==3.8.0
typer==0.9.0
types-cachetools==5.3.0.6
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.7.12
uptrace==1.18.0
uritemplate==4.1.1
urllib3==1.26.16
uuid==1.30
uvicorn==0.22.0
uvloop==0.17.0
validators==0.20.0
virtualenv==20.23.0
watchfiles==0.19.0
weaviate-client==3.22.1
webencodings==0.5.1
websocket-client==1.6.1
websockets==10.4
Werkzeug==2.3.4
wikipedia==1.4.0
wrapt==1.15.0
xlrd==2.0.1
XlsxWriter==3.1.2
yarl==1.9.2
zipp==3.15.0
zstandard==0.21.0
